code restructure + SOC
