--- frontend/src/components/TestPush.js
@@ -1,1 +1,0 @@
-aW1wb3J0IFJlYWN0IGZyb20gJ3JlYWN0JzsKCmNvbnN0IFRlc3RQdXNoID0gKCkgPT4gewogIHJldHVybiAoCiAgICA8ZGl2IGNsYXNzTmFtZT0idGV4dC1jZW50ZXIgcC00Ij4KICAgICAgPGgxPlRlc3QgUHVzaCBTdWNjZXNzZnVsITwvaDE+CiAgICA8L2Rpdj4KICB9Owp9OwoKZXhwb3J0IGRlZmF1bHQgVGVzdFB1c2g7
